moved value renderer to utils
